NLP/requirements.txt
flask==2.0.3
imdbpy==2021.4.18
google-cloud-language==2.3.1
redis==4.2.2
waiting